% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_server.R
\name{create_server}
\alias{create_server}
\title{Function provides an http server that connects R language resources with javascript applications.}
\usage{
create_server(
  index_path = NULL,
  port = 8080,
  static_paths = NULL,
  routes = NULL
)
}
\arguments{
\item{index_path}{A string that defines the full directory path to an \code{index.html} file defining the server's
root path (i.e. "\"). The root html file is assumed to be named "index.html".}

\item{port}{An integer that defines the port number for the server. The default is 8080.}

\item{static_paths}{A named list where the string name references the web route (e.g. a web path "/assets", "/styles", "/html")
and the corresponding string value references a \strong{full} static directory path. The server will host static
resources (css, html, etc) in these directories.}

\item{routes}{A named list where the string name references specific web routes and the value references an R callback function
that optionally returns a JSON formatted string for the response body. The callback should have a list argument where it will receive
information on the request including the request method, body, type, and query string upon which it can act on.}
}
\value{
A \code{httpuv::WebServer}
}
\description{
The package is based on \href{https://cran.r-project.org/web/packages/httpuv/index.html}{httpuv::}.
 It is designed to host static resources such as html, javascript, css, etc along with specific route requests
 to perform various R based actions or respond with JSON formatted content.  The server is designed to work in a local
 host environment as part of an html/javascript project that needs a link to R language capabilities and its
 associated packages. As per the \code{httpuv::} documentation, the server can be stopped by calling
 \code{httpuv::stopServer(server=name_of_server)}. Also to list the current running server(s) call \code{httpuv::listServers()}.
}
\author{
Rick Dean
}
